feat(build): add master script and helper scripts to deploy app on server
Description: attempting to make use of the CRA recommended static deployment
server for the frontend. New script + helpers launches api server, builds ui
project using CRA scripts, and deploys compiled code using serve.

Note: this still doesn't seem to be getting us what we want, as the compiled code
seems to have javascript errors, but it's possible that this is a source code issue
and not a deploy issue. Pushing this so it will get to our test server, which will also
have the side effect with the current pipeline of pushing the compiled UI to the static
s3 server, we can see then if it has the same issue.