fix(build): remove unused deploy scripts and hardcoded auth endpoints
Description: the yarn scripts for deployment that I added yesterday
proved unnecessary to the deployment solution that ended up working.
Also, some of the endpoints for Google auth were hardcoded to
localhost, we can simply remove the hostname from these as they will
be caught by the proxy in development and by nginx routing in the
deployed environments.