feat(api): add getEventTemplate and editEventTemplate to api
Description: getEventTemplate fulfills a GET request, returning the
relevant event template info to prefill the edit event template
form. editEventTemplate fulfills a POST request, updating the
event template in the database with the values from a submitted
edit event template form.

Note: consider this a WIP until the edit event template form is finished
and checked in, after which this can be tested and adjusted as necessary.