fix(build): remove .env file from repo and gitignore, add example
Description: the .env file under api should probably never have
been in our repo, as this is more of a local config thing, and
even though the passwords in our dev environment don't matter,
it's bad practice to be storing that stuff upstream. Also, having
it there makes things more complicated for our test environment,
which needs to have its own .env file which should *definitely*
not be distributed. This change removes the .env file from our
repo and adds it to our gitignore, and adds in a .env.example
file strictly so that future contributors can get an idea
of what's needed in their *local* .env file.