feat(build): move npm packages from top level to package/api
Description: These packages were added at the top level, but they really belong in
packages/api. This doesn't cause any problems in dev, but when just the files
that we ought to need for the webserver are transferred up (as opposed to cloning
the entire project), we end up missing packages that are needed by the api at
runtime. I suspect that there are a few packages remaining in the top-level that
should really be in packages/ui, but these aren't as important, because the ui
files that end up on the webserver are compiled.